services/worker-service/requirements.txt
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
boto3==1.34.0
requests==2.31.0
google-api-python-client==2.108.0
google-auth==2.25.0
python-dotenv==1.0.0
dropbox==11.36.0